Smol fix + refactor and clean up
